fix delete home

+++ src/main/java/com/hotel/demo/controller/ApiHomeController.java
@@ -1,7 +1,9 @@
 package com.hotel.demo.controller;
 
 import com.hotel.demo.message.request.SearchByAddress;
+import com.hotel.demo.model.Comment;
 import com.hotel.demo.model.Home;
+import com.hotel.demo.service.CommentService;
 import com.hotel.demo.service.HomeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -17,6 +19,9 @@ import java.util.Optional;
 public class ApiHomeController {
     @Autowired
     private HomeService homeService;
+
+    @Autowired
+    private CommentService commentService;
 
     @GetMapping
     public ResponseEntity<List<Home>> findAllHome() {
@@ -86,11 +91,15 @@ public class ApiHomeController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Home> deleteHome(@PathVariable Long id) {
+    public ResponseEntity<?> deleteHome(@PathVariable Long id) {
+        List<Comment> comments = (List<Comment>) commentService.findCommentByHomeId(id);
         Optional<Home> home = homeService.findById(id);
-        if (home.isPresent()) {
+        if (!comments.isEmpty() && home.isPresent()) {
+            for (Comment comment : comments) {
+                commentService.remove(comment.getId());
+            }
             homeService.remove(id);
-            return new ResponseEntity<>(home.get(), HttpStatus.OK);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
